+++ src/pages/donation-calculator/pdf-export/data.json
@@ -1,0 +1,220 @@
+{
+  "amount": 100000,
+  "ogMissedFromDonTypes": 18500,
+  "ogFees": 4900,
+  "ogSubsCost": 1200,
+  "ogDeductions": 6100,
+  "ogNet": 93900,
+  "bgFees": 399.9999999999999,
+  "bgNet": 119300,
+  "feeSavings": 4500,
+  "advantage": 24200,
+  "notGranted": 11930,
+  "notGrantedRate": 0.1,
+  "savings": 5965,
+  "savingsRate": 0.5,
+  "invested": 5965,
+  "investedRate": 0.5,
+  "projection": [
+    {
+      "end": {
+        "liq": 6208.422661534866,
+        "lock": 7285.2683944717965,
+        "total": 13493.691056006663
+      },
+      "liq": 243.42266153486617,
+      "lock": 1320.2683944717965,
+      "total": 1563.6910560066626
+    },
+    {
+      "end": {
+        "liq": 12670.201696616103,
+        "lock": 16183.027921628496,
+        "total": 28853.2296182446
+      },
+      "liq": 740.2016966161027,
+      "lock": 4253.027921628496,
+      "total": 4993.229618244599
+    },
+    {
+      "end": {
+        "liq": 19395.676197073637,
+        "lock": 27050.180690246238,
+        "total": 46445.856887319875
+      },
+      "liq": 1500.676197073637,
+      "lock": 9155.180690246238,
+      "total": 10655.856887319875
+    },
+    {
+      "end": {
+        "liq": 26395.607177495356,
+        "lock": 40322.624043659045,
+        "total": 66718.2312211544
+      },
+      "liq": 2535.607177495356,
+      "lock": 16462.624043659045,
+      "total": 18998.2312211544
+    },
+    {
+      "end": {
+        "liq": 33681.194793259005,
+        "lock": 56532.73503779747,
+        "total": 90213.92983105648
+      },
+      "liq": 3856.1947932590047,
+      "lock": 26707.735037797473,
+      "total": 30563.929831056477
+    },
+    {
+      "end": {
+        "liq": 41264.09626120617,
+        "lock": 76330.7248611799,
+        "total": 117594.82112238607
+      },
+      "liq": 5474.09626120617,
+      "lock": 40540.7248611799,
+      "total": 46014.82112238607
+    },
+    {
+      "end": {
+        "liq": 49156.444511631824,
+        "lock": 100510.71975378049,
+        "total": 149667.1642654123
+      },
+      "liq": 7401.444511631824,
+      "lock": 58755.71975378049,
+      "total": 66157.16426541231
+    },
+    {
+      "end": {
+        "liq": 57370.86760143364,
+        "lock": 130042.61456845807,
+        "total": 187413.4821698917
+      },
+      "liq": 9650.867601433638,
+      "lock": 82322.61456845807,
+      "total": 91973.48216989171
+    },
+    {
+      "end": {
+        "liq": 65920.50891948237,
+        "lock": 166110.97666774315,
+        "total": 232031.48558722553
+      },
+      "liq": 12235.508919482367,
+      "lock": 112425.97666774315,
+      "total": 124661.48558722551
+    },
+    {
+      "end": {
+        "liq": 74819.04821654345,
+        "lock": 210162.560648009,
+        "total": 284981.60886455246
+      },
+      "liq": 15169.048216543451,
+      "lock": 150512.560648009,
+      "total": 165681.60886455246
+    },
+    {
+      "end": {
+        "liq": 84080.72349339709,
+        "lock": 263964.33977589716,
+        "total": 348045.0632692942
+      },
+      "liq": 18465.72349339709,
+      "lock": 198349.33977589716,
+      "total": 216815.06326929425
+    },
+    {
+      "end": {
+        "liq": 93720.353782181,
+        "lock": 329674.38186252804,
+        "total": 423394.735644709
+      },
+      "liq": 22140.353782181002,
+      "lock": 258094.38186252804,
+      "total": 280234.735644709
+    },
+    {
+      "end": {
+        "liq": 103753.36285740347,
+        "lock": 409928.41251016274,
+        "total": 513681.7753675662
+      },
+      "liq": 26208.362857403466,
+      "lock": 332383.41251016274,
+      "total": 358591.7753675662
+    },
+    {
+      "end": {
+        "liq": 114195.80391456984,
+        "lock": 507945.537909355,
+        "total": 622141.3418239248
+      },
+      "liq": 30685.803914569842,
+      "lock": 424435.537909355,
+      "total": 455121.34182392486
+    },
+    {
+      "end": {
+        "liq": 125064.38525590315,
+        "lock": 627657.3678821509,
+        "total": 752721.753138054
+      },
+      "liq": 35589.38525590315,
+      "lock": 538182.3678821509,
+      "total": 573771.753138054
+    },
+    {
+      "end": {
+        "liq": 136376.49702426308,
+        "lock": 773865.7184848636,
+        "total": 910242.2155091267
+      },
+      "liq": 40936.49702426308,
+      "lock": 678425.7184848636,
+      "total": 719362.2155091267
+    },
+    {
+      "end": {
+        "liq": 148150.23902803552,
+        "lock": 952435.2198518036,
+        "total": 1100585.4588798392
+      },
+      "liq": 46745.23902803552,
+      "lock": 851030.2198518036,
+      "total": 897775.4588798392
+    },
+    {
+      "end": {
+        "liq": 160404.44970151185,
+        "lock": 1170528.555061385,
+        "total": 1330933.0047628968
+      },
+      "liq": 53034.449701511854,
+      "lock": 1063158.555061385,
+      "total": 1116193.0047628968
+    },
+    {
+      "end": {
+        "liq": 173158.73624710014,
+        "lock": 1436893.7657985669,
+        "total": 1610052.502045667
+      },
+      "liq": 59823.73624710014,
+      "lock": 1323558.7657985669,
+      "total": 1383382.502045667
+    },
+    {
+      "end": {
+        "liq": 186433.50600759257,
+        "lock": 1762215.1490626778,
+        "total": 1948648.6550702704
+      },
+      "liq": 67133.50600759257,
+      "lock": 1642915.1490626778,
+      "total": 1710048.6550702704
+    }
+  ]
+}
